pyproject.toml
[tool.poetry]
name = "noaa"
version = "0.1.0"
description = ""
authors = ["Mushinako <github@mushinako.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.12.0"
ExifRead = "^3.0.0"
GeoAlchemy2 = { extras = ["shapely"], version = "^0.14.1" }
PyYAML = "^6.0.1"
SQLAlchemy = "^2.0.20"

[tool.poetry.group.dev.dependencies]
black = "^23.9.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
